app/requirements.txt
#requests==2.32.3
#gradio==5.23.3
#gradio_client==1.8.0
#fastapi
#uvicorn
plotly==6.0.1
matplotlib==3.7.5
matplotlib-inline==0.1.7
pandas==2.2.3
databricks-sdk
folium
streamlit
branca
